Commands

Typstry

Typstry.TypstCommand — Type

TypstCommand(::Vector{String})
TypstCommand(::TypstCommand; kwargs...)
TypstCommand(::TypstCommand, ignorestatus, flags, env, dir, cpus = nothing)

The Typst compiler.

Info

This type implements the Cmd interface. However, the interface is unspecified, which may result in unexpected behavior.

Examples

julia> help = TypstCommand(["help"])
typst`help`

julia> TypstCommand(help; ignorestatus = true)
typst`help`

Typstry.TypstError — Type

TypstError <: Exception
TypstError(::TypstCommand)

An Exception indicating a failure to run a TypstCommand.

Typstry.@typst_cmd — Macro

@typst_cmd(s)
typst`s`

Construct a TypstCommand where each parameter is separated by a space.

This does not support interpolation; use the constructor instead.

Examples

julia> typst`help`
typst`help`

julia> typst`compile input.typ output.typ`
typst`compile input.typ output.typ`

Typstry.julia_mono — Constant

julia_mono

An constant artifact containing the JuliaMono typeface.

Use with a TypstCommand and one of addenv, setenv, or the font-path Typst command-line option.

Typstry.render — Function

render(x;
    input = "input.typ",
    output = "output.pdf",
    open = true,
    preamble = true,
context...)

Render to a document using show(::IO, ::MIME"text/typst", ::Typst).

This generates two files: the input is the Typst source text and the output is the compiled document. The document format is inferred by the file extension of output, which may be pdf, png, or svg. The document may be automatically opened by the default viewer. The Typst source text may begin with the following preamble:

#set page(margin: 1em, height: auto, width: auto, fill: white)
#set text(16pt, font: "JuliaMono")

Examples

julia> render(Any[true 1; 1.2 1 // 2]);

Base

Base.:== — Function

==(::TypstCommand, ::TypstCommand)

See also TypstCommand.

Examples

julia> help = typst`help`
typst`help`

julia> help == help
true

julia> help == TypstCommand(help; ignorestatus = true)
false

Base.addenv — Function

addenv(::TypstCommand, args...; kwargs...)

See also TypstCommand and julia_mono.

Examples

julia> addenv(typst`compile input.typ output.pdf`, "TYPST_FONT_PATHS" => julia_mono)
typst`compile input.typ output.pdf`

Base.detach — Function

detach(::TypstCommand)

See also TypstCommand.

Examples

julia> detach(typst`help`)
typst`help`

Base.eltype — Function

eltype(::Type{TypstCommand})

See also TypstCommand.

Examples

julia> eltype(TypstCommand)
String

Base.firstindex — Function

firstindex(::TypstCommand)

See also TypstCommand.

Examples

julia> firstindex(typst`help`)
1

Base.getindex — Function

getindex(::TypstCommand, i)

See also TypstCommand.

Examples

julia> typst`help`[2]
"help"

Base.hash — Function

hash(::TypstCommand, ::UInt)

See also TypstCommand.

Base.ignorestatus — Function

ignorestatus(::TypstCommand)

Return a TypstCommand that does not throw a TypstError if the Typst compiler throws an error.

Errors thrown by the Typst compiler are printed to stderr regardless.

Examples

julia> ignorestatus(typst`help`)
typst`help`

Base.iterate — Method

iterate(::TypstCommand)
iterate(::TypstCommand, i)

See also TypstCommand.

Examples

julia> iterate(typst`help`, 2)
("help", 3)

julia> iterate(typst`help`, 3)

Base.keys — Function

keys(::TypstCommand)

See also TypstCommand.

Examples

julia> keys(typst`help`)
2-element LinearIndices{1, Tuple{Base.OneTo{Int64}}}:
 1
 2

Base.lastindex — Function

lastindex(::TypstCommand)

See also TypstCommand.

Examples

julia> lastindex(typst`help`)
2

Base.length — Function

length(::TypstCommand)

See also TypstCommand.

Examples

julia> length(typst`help`)
2

Base.run — Function

run(::TypstCommand, args...; kwargs...)

See also TypstCommand.

Info

If the Typst compiler throws an error, it will be printed to stderr. Then, a Julia TypstError will be thrown unless the ignorestatus flag is set.

Examples

julia> run(typst`compile input.typ output.pdf`);

Base.setcpuaffinity — Function

setcpuaffinity(::TypstCommand, cpus)

See also TypstCommand.

Compat

Requires at least Julia v0.8.

Examples

julia> setcpuaffinity(typst`help`, nothing)
typst`help`

Base.setenv — Function

setenv(::TypstCommand, args...; kwargs...)

See also TypstCommand and julia_mono.

Examples

julia> setenv(typst`compile input.typ output.pdf`, "TYPST_FONT_PATHS" => julia_mono)
typst`compile input.typ output.pdf`

Base.show — Method

show(::IO, ::MIME"text/plain", ::TypstCommand)

See also TypstCommand.

Examples

julia> show(stdout, "text/plain", typst`help`)
typst`help`

Base.showerror — Function

showerror(::IO, ::TypstError)

Print a TypstError when failing to run a TypstCommand.

Examples

julia> showerror(stdout, TypstError(typst``))
TypstError: failed to `run` a `TypstCommand([""]))`