Internals

Typstry.@stable_disable — Macro

@stable_disable(x)

Equivalent to DispatchDoctor.@stable default_mode = "disable" x.

Typstry.workload — Function

workload(examples)

Strings

Typstry.examples — Constant

examples

A constant Vector of Julia values and their corresponding Types implemented for show_typst.

Typstry.preamble — Constant

preamble

A constant used at the beginning of Typst source files generated by show(::IO, ::Union{MIME"application/pdf", MIME"image/png", MIME"image/svg+xml"}, ::TypstString).

Examples

julia> println(Typstry.preamble)
#set page(margin: 1em, height: auto, width: auto, fill: white)
#set text(16pt, font: "JuliaMono")

Typstry.typst_mime — Constant

typst_mime

Examples

julia> Typstry.typst_mime
MIME type text/typst

Typstry.backticks — Function

backticks(io)

Return to io[:backticks]::Int.

Examples

julia> Typstry.backticks(IOContext(stdout, :backticks => 3))
3

Typstry.block — Function

block(io)

Return io[:block]::Bool.

Examples

julia> Typstry.block(IOContext(stdout, :block => true))
true

Typstry.code_mode — Function

code_mode(io)

Print the number sign, unless mode(io) == code.

See also Mode and mode.

Examples

julia> Typstry.code_mode(IOContext(stdout, :mode => code))

julia> Typstry.code_mode(IOContext(stdout, :mode => markup))
#

julia> Typstry.code_mode(IOContext(stdout, :mode => math))
#

Typstry.depth — Function

depth(io)

Return io[:depth]::Int.

Examples

julia> Typstry.depth(IOContext(stdout, :depth => 0))
0

Typstry.enclose — Function

enclose(f, io, x, left, right = reverse(left); kwargs...)

Call f(io, x; kwargs...) between printing left and right, respectfully.

Examples

julia> Typstry.enclose((io, i; x) -> print(io, i, x), stdout, 1, "\$ "; x = "x")
$ 1x $

Typstry.format — Function

format(::Union{MIME"application/pdf", MIME"image/png", MIME"image/svg+xml"})

Examples

julia> Typstry.format(MIME"application/pdf"())
"pdf"

julia> Typstry.format(MIME"image/png"())
"png"

julia> Typstry.format(MIME"image/svg+xml"())
"svg"

Typstry.indent — Function

indent(io)

Return " " ^ io[:tab_size]::Int.

See also TypstString.

Examples

julia> Typstry.indent(IOContext(stdout, :tab_size => 2))
"  "

Typstry.join_with — Function

join_with(f, io, xs, delimeter; kwargs...)

Similar to join, except printing with f(io, x; kwargs...).

Examples

julia> Typstry.join_with((io, i; x) -> print(io, -i, x), stdout, 1:4, ", "; x = "x")
-1x, -2x, -3x, -4x

Typstry.math_pad — Function

math_pad(io, x)

Return "", "\$", or "\$ " depending on the block and mode settings.

Examples

julia> Typstry.math_pad(IOContext(stdout, :mode => math))
""

julia> Typstry.math_pad(IOContext(stdout, :block => true, :mode => markup))
"\$ "

julia> Typstry.math_pad(IOContext(stdout, :block => false, :mode => markup))
"\$"

Typstry.maybe_wrap — Function

maybe_wrap(::Any)

Wrap the value in Typst unless it is a TypstString or TypstText.

Examples

julia> Typstry.maybe_wrap(1)
Typst{Int64}(1)

julia> Typstry.maybe_wrap(TypstString(1))
typst"#1"

julia> Typstry.maybe_wrap(TypstText(1))
TypstText{Int64}(1)

Typstry.mode — Function

mode(io)

Return io[:mode]::Mode.

See also Mode.

Examples

julia> Typstry.mode(IOContext(stdout, :mode => code))
code::Mode = 0

Typstry.parenthesize — Function

parenthesize(io)

Return io[:parenthesize]::Bool.

Examples

julia> Typstry.parenthesize(IOContext(stdout, :parenthesize => true))
true

Typstry.print_parameters — Function

print_parameters(io, f, keys, final)

Examples

julia> Typstry.print_parameters(
           IOContext(stdout, :depth => 0, :tab_size => 2, :delim => typst"\"(\""),
       "vec", [:delim, :gap], true)
vec(
  delim: "(",

Typstry.show_array — Function

show_array(io, x)

Typstry.show_raw — Function

show_raw(io, x)

Typstry.show_vector — Function

show_vector(io, x)

Typstry.static_parse — Function

static_parse(args...; filename, kwargs...)

Call Meta.parse with the filename if it is supported in the current Julia version (at least v1.10).

Commands

Typstry.typst_compiler — Constant

typst_compiler

A constant Cmd that is the Typst compiler given by Typst_jll.jl with no additional parameters.

Typstry.apply — Function

apply(f, tc, args...; kwargs...)

Extensions

Dates