Strings

Typstry

Typstry.Mode — Type

Mode

An Enumerated type used to specify that the current Typst context is in code, markup, or math mode.

julia> Mode
Enum Mode:
code = 0
markup = 1
math = 2

Typstry.Typst — Type

Typst{T}
Typst(::T)

A wrapper used to pass values to show(::IO, ::MIME"text/typst", ::Union{Typst, TypstString}).

julia> Typst(1)
Typst{Int64}(1)

julia> Typst("a")
Typst{String}("a")

Typstry.TypstString — Type

TypstString <: AbstractString
TypstString(::Any; context...)

Convert the value to a Typst formatted string.

Optional Julia settings and Typst parameters are passed to show(::IO, ::MIME"text/typst", ::Union{Typst, TypstString}) in an IOContext. See also show_typst for a list of supported types.

Info

This type implements the String interface. However, the interface is unspecified which may result unexpected behavior.

Examples

julia> TypstString("a")
typst"\"a\""

julia> TypstString("a"; mode = code)
typst"\"\\\"a\\\"\""

Typstry.@typst_str — Macro

@typst_str(s)
typst"s"

Construct a TypstString.

Control characters are escaped, except quotation marks and backslashes in the same manner as @raw_str. TypstStrings containing control characters may be created using typst_text. Values may be interpolated by calling the TypstString constructor, except using a backslash instead of the type name.

Tip

Pring directly to an IO using show(::IO, ::MIME"text/typst", ::Union{Typst, TypstString}).

See also the performance tip to Avoid string interpolation for I/O.

Examples

julia> x = 1;

julia> typst"$ \(x) / \(x + 1) $"
typst"$ 1 / 2 $"

julia> typst"\(x // 2)"
typst"$1 / 2$"

julia> typst"\(x // 2; mode = math)"
typst"(1 / 2)"

julia> typst"\\(x)"
typst"\\(x)"

Typstry.show_typst — Function

show_typst(io, x)

Print to Typst format using Julia settings and Typst parameters in the IOContext.

Settings are used in Julia to format the TypstString and can be any type. Parameters are passed to a function in the Typst source file and must be a String with the same name as in Typst, except that dashes are replaced with underscores.

For additional information on parameters and settings, see also show(::IO, ::MIME"text/typst", ::Union{Typst, TypstString}) and the Typst Documentation, respectively.

For additional information on printing and rendering, see also Examples.

Tip

Implement this function for custom types to specify their Typst formatting.

Warning

This function's methods are incomplete. Please file an issue or create a pull-request for missing methods.

Type	Settings	Parameters

AbstractChar	:mode

AbstractFloat

AbstractMatrix	:block, :depth, :indent, :mode	:augment, :column_gap, :delim, :gap, :row_gap

AbstractString	:mode

AbstractVector	:block, :depth, :indent, :mode	:delim, :gap

Bool	:mode

Complex	:block, :mode, :parenthesize

Irrational	:mode

Nothing	:mode

OrdinalRange{<:Integer, <:Integer}	:mode

Rational	:block, :mode, :parenthesize

Regex	:mode

Signed

StepRangeLen{<:Integer, <:Integer, <:Integer}	:mode

Text	:mode

Typst	...	...

TypstString

Examples

julia> show_typst(stdout, 1)
1

julia> show_typst(IOContext(stdout, :mode => code), "a")
"\"a\""

Typstry.typst_text — Function

typst_text(::Any)

Construct a TypstString using print instead of show(::IO, ::MIME"text/typst", ::Union{Typst, TypstString}).

Warning

Unescaped control characters in TypstStrings may break formatting in some environments, such as the REPL.

Examples

julia> typst_text("a")
typst"a"

julia> typst_text([1, 2, 3, 4])
typst"[1, 2, 3, 4]"

Base

Base.IOBuffer — Type

IOBuffer(::TypstString)

See also TypstString.

Examples

julia> IOBuffer(typst"a")
IOBuffer(data=UInt8[...], readable=true, writable=false, seekable=true, append=false, size=1, maxsize=Inf, ptr=1, mark=-1)

Base.codeunit — Function

codeunit(::TypstString)
codeunit(::TypstString, ::Integer)

See also TypstString.

Examples

julia> codeunit(typst"a")
UInt8

julia> codeunit(typst"a", 1)
0x61

Base.isvalid — Function

isvalid(::TypstString, ::Integer)

See also TypstString.

Examples

julia> isvalid(typst"a", 1)
true

Base.iterate — Method

iterate(::TypstString, ::Integer)
iterate(::TypstString)

See also TypstString.

Examples

julia> iterate(typst"a")
('a', 2)

julia> iterate(typst"a", 1)
('a', 2)

Base.ncodeunits — Function

ncodeunits(::TypstString)

See also TypstString.

Examples

julia> ncodeunits(typst"a")
1

Base.pointer — Function

pointer(::TypstString)

See also TypstString.

Base.repr — Function

repr(::MIME, ::TypstString; kwargs...)

See also TypstString.

Info

This method patches incorrect output from the assumption in repr that the parameter is already in the requested MIME type when the MIME type istextmime and the parameter is an AbstractString.

Examples

julia> repr(MIME"text/plain"(), typst"a")
"typst\"a\""

julia> repr(MIME"text/typst"(), typst"a")
typst"a"

Base.show — Method

show(::IO, ::TypstString)

See also TypstString.

Examples

julia> show(stdout, typst"a")
typst"a"

Base.show — Method

show(::IO, ::MIME"text/typst", ::Union{Typst, TypstString})

Print to Typst format.

Provides default settings for show_typst which may be specified in an IOContext.

See also Typst and TypstString.

Tip

Implement this function for custom types to specify their default settings and parameters.

Setting	Default	Type	Description

:block	false	Bool	When :mode => math, specifies whether the enclosing dollar signs are padded with a space to render the element inline or its own block.

:depth	0	Int	The current level of nesting within container types to specify the degree of indentation.

:indent	' ' ^ 4	String	The string used for horizontal spacing by some elements with multi-line Typst formatting.

:mode	markup	Mode	The current Typst context where code follows the number sign, markup is at the top-level and enclosed in square brackets, and math is enclosed in dollar signs.

:parenthesize	true	Bool	Whether to enclose some mathematical elements in parentheses.

Examples

julia> show(stdout, "text/typst", typst"a")
a

julia> show(stdout, "text/typst", Typst("a"))
"a"

julia> show(IOContext(stdout, :mode => code), "text/typst", Typst("a"))
"\"a\""

Base.show — Method

show(::IO, ::Union{
    MIME"application/pdf", MIME"image/png", MIME"image/svg+xml"
}, ::TypstString)

Print to Portable Document Format (PDF), Portable Network Graphics (PNG), or Scalable Vector Graphics (SVG) format.

The corresponding Typst source file begins with this preamble:

#set page(margin: 1em, height: auto, width: auto, fill: white)
#set text(16pt, font: "JuliaMono")

Note

Environments such as Pluto.jl notebooks use this function to render TypstStrings to a document.