Render

Typstry.render — Function

render(::TypstContext, value; parameters...)
render(value; parameters..., context...)

Render the value to a document.

This supports using the julia_mono typeface.

See also TypstContext.

Info

Typst requires the output path to contain a page number template {p} when compiling documents with multiple pages to either PNG or SVG format. In this case, use open = false to prevent throwing a TypstCommandError. See also Typst #7182.

Parameters

input::AbstractString = "input.typ"

Write the preamble and formatted value to this Typst source file.

output::AbstractString = "output.pdf"

Compile the document in the format specified by the file extension pdf, png, or svg.

open::Bool = true

Whether to preview the document with the default viewer, if available.

ignorestatus::Bool = true

Whether to throw a TypstCommandError if the command errors.

Examples

julia> render(Any[true 1; 1.2 1 // 2]);