Strings

This reference documents formatting Julia values into Typst source text.

Typstry.TypstString — Type

TypstString <: AbstractString
TypstString(::TypstContext, ::Any)
TypstString(::Any; context...)

Format the value as a Typst formatted string.

The TypstContext is combined with additional context and passed to show_typst.

Interface

This type implements the String interface. However, the interface is undocumented, which may result in unexpected behavior.

IOBuffer(::TypstString)

codeunit(::TypstString, ::Integer)

codeunit(::TypstString)

isvalid(::TypstString, ::Integer)

iterate(::TypstString, ::Integer)

iterate(::TypstString)

ncodeunits(::TypstString)

pointer(::TypstString)

repr(::MIME, ::TypstString)

This method patches incorrect output from the assumption in repr that the parameter is already in the requested MIME type when the MIME type satisfies istextmime and the parameter is an AbstractString.

show_typst(::IO, ::TypstContext, ::TypstString)

show(::IO, ::MIME"text/typst", ::TypstString)

Accepts a IOContext(::IO, :typst_context => ::TypstContext)

show(::IO, ::Union{MIME"application/pdf", MIME"image/png", MIME"image/svg+xml"}, ::TypstString)

Accepts a IOContext(::IO, :typst_context => ::TypstContext)

Supports the julia_mono typeface

The generated Typst source text contains the context's preamble and the formatted value

show(::IO, ::TypstString)

Prints in @typst_str format if each character satisfies isprint. Otherwise, print in TypstString format.

Examples

julia> TypstString(1)
typst"$1$"

julia> TypstString(TypstContext(; mode = code), π)
typst"3.141592653589793"

julia> TypstString(1 + 2im; mode = math)
typst"(1 + 2i)"

Typstry.TypstText — Type

TypstText{T}
TypstText(::Any)

A wrapper whose show_typst method uses print on the wrapped value.

Interface

show_typst(::IO, ::TypstContext, ::TypstText)

show(::IO, ::MIME"text/typst", ::TypstText)

Accepts IOContext(::IO, :typst_context => ::TypstContext)

show(::IO, ::Union{MIME"application/pdf", MIME"image/png", MIME"image/svg+xml"}, ::TypstText)

Accepts IOContext(::IO, :typst_context => ::TypstContext)

Supports the julia_mono typeface

The generated Typst source text contains the context's preamble and the formatted value

show(::IO, ::TypstText)

Examples

julia> TypstText(1)
TypstText(1)

julia> TypstText("a")
TypstText("a")

Typstry.Typst — Type

Typst{T}
Typst(::T)

A wrapper used to pass values to show_typst.

Interface

show_typst(::IO, ::TypstContext, ::Typst)

show(::IO, ::Union{MIME"application/pdf", MIME"image/png", MIME"image/svg+xml", MIME"text/typst"}, ::Typst)

Accepts a IOContext(::IO, :typst_context => ::TypstContext)

Supports the julia_mono typeface

The generated Typst source text contains the context's preamble and the formatted value

show(::IO, ::Typst)

Examples

julia> Typst(1)
Typst(1)

julia> Typst("a")
Typst("a")

Typstry.@typst_str — Macro

@typst_str("s")
typst"s"

Construct a TypstString.

Control characters are escaped, except double quotation marks and backslashes in the same manner as @raw_str. Values may be interpolated by calling the TypstString constructor, except using a backslash instead of the type name. Interpolation syntax may be escaped in the same manner as quotation marks.

Tip

Print directly to an IO using show(::IO, ::MIME"text/typst", ::Typst).

See also the performance tip to Avoid string interpolation for I/O.

Examples

julia> x = 1;

julia> typst"$ \(x; mode = math) / \(x + 1; mode = math) $"
typst"$ 1 / 2 $"

julia> typst"\(x//2)"
typst"$1 / 2$"

julia> typst"\(x // 2; mode = math)"
typst"(1 / 2)"

julia> typst"\\(x)"
typst"\\(x)"

Typstry.show_typst — Function

show_typst(::IO = stdout, ::TypstContext, x)
show_typst(::TypstContext = TypstContext(), x)

Print in Typst format with Julia settings and Typst parameters provided by the TypstContext.

Implement the three-parameter form of this function for a custom type to specify its Typst formatting. A setting is a value used in Julia, whose type varies across settings. A parameter is passed directly to a Typst function and must be a TypstString with the same name as in Typst, except that dashes are replaced with underscores. Settings each have a default value specified by context, whereas the default values of parameters are handled in Typst functions. Some settings, such as block, correspond with a parameter but may also be used in Julia.

Tip

Please create an issue or pull-request to implement new methods.

Type	Settings	Parameters

AbstractArray	:block, :depth, :mode, :tab_size	:delim, :gap

AbstractChar

AbstractFloat	:mode

AbstractMatrix	:block, :depth, :mode, :tab_size	:augment, :column_gap, :delim, :gap, :row_gap

AbstractString

Bool	:mode

Complex{Bool}	:block, :mode, :parenthesize

Complex	:block, :mode, :parenthesize

Irrational	:mode

Nothing	:mode

OrdinalRange{<:Integer, <:Integer}	:mode

Rational	:block, :mode, :parenthesize

Regex	:mode

Signed	:mode

StepRangeLen{<:Integer, <:Integer, <:Integer}	:mode

Tuple	:block, :depth, :mode, :tab_size	:delim, :gap

Typst

TypstString

TypstText

Unsigned	:mode

VersionNumber	:mode

Docs.HTML	:block, :depth, :mode, :tab_size

Docs.Text	:mode

Dates.Date	:mode, :indent

Dates.DateTime	:mode, :indent

Dates.Period	:mode, :indent

Dates.Time	:mode, :indent

Typstry.show_typst — Method

show_typst(::IO, ::TypstContext, ::TypstString)

Typstry.show_typst — Method

show_typst(::IO, ::TypstContext, ::TypstText)

Call print the value wrapped in TypstText.

See also TypstContext.

Typstry.show_typst — Method

show_typst(::IO, ::TypstContext, ::Typst)

Call show_typst on the value wrapped in Typst.

See also TypstContext.