Utilities

Typstry.Mode — Type

Mode

An Enumerated type used to specify that the current Typst syntactical context is code, markup, or math.

Examples

julia> Mode
Enum Mode:
code = 0
markup = 1
math = 2

Typstry.code — Constant

code

A Typst syntactical Mode prefixed by the number sign.

Examples

julia> code
code::Mode = 0

Typstry.markup — Constant

markup

A Typst syntactical Mode at the top-level of source text and enclosed within square brackets.

julia> markup
markup::Mode = 1

Typstry.math — Constant

math

A Typst syntactical Mode enclosed within dollar signs.

julia> math
math::Mode = 2

Typstry.ContextError — Type

ContextError <: Exception
ContextError(::Type, ::Type, ::Symbol)

An Exception indicating that a context key returned a value of an incorrect type.

Interface

Implements the Exception interface.

showerror(::IO, ::ContextError)

show(::IO, ::MIME"text/plain", ::ContextError)

Examples

julia> ContextError(Mode, String, :mode)
ContextError(Mode, String, :mode)

Typstry.TypstContext — Type

TypstContext <: AbstractDict{Symbol, Any}
TypstContext(::Any)
TypstContext(; kwargs...)

Provide formatting data for show_typst.

Implement a method of this constructor for a custom type to specify its custom settings and parameters.

Calls to show_typst from the following methods:

TypstString

render

show_typst(::IO, ::TypstContext, x)

show with the application/pdf, image/png, image/svg+xml, and text/typst MIME types and a TypstString, TypstText, and Typst value

specify the TypstContext by combining the following contexts:

The context

Any context specified by implementing the TypstContext constructor for the given type

The context specified in the call by keyword parameters, a given TypstContext, or the IOContext key :typst_context, depending on the calling method

Any context specified by a recursive call in show_typst to format values, such as elements from a container

Duplicate keys are handled such that each successive context overwrites those of previous contexts, prioritized in order as listed. In other words, the default context has the lowest priority while recursive calls to show_typst have the highest priority.

Interfaces

This type implements the dictionary and iteration interfaces. However, it does not support removing mappings.

copy(::TypstContext)

eltype(::TypstContext)

getkey(::TypstContext, ::Any, ::Any)

get(::TypstContext, ::Any, ::Any)

get(::Union{Function, Type}, ::TypstContext, ::Any)

iterate(::TypstContext, state)

iterate(::TypstContext)

length(::TypstContext)

mergewith(::Any, ::TypstContext, ::AbstractDict...)

merge!(::TypstContext, ::AbstractDict...)

merge(::TypstContext, ::AbstractDict...)

setindex!(::TypstContext, ::Any, ::Any)

show(::IO, ::TypstContext)

sizehint!(::TypstContext, ::Any)

Typstry.context — Constant

context

A constant TypstContext used to provide default formatting data when calling show_typst.

See also reset_context.

Tip

Set mappings in this dictionary to customize the default formatting in environments that display values using show with the application/pdf, image/png, and image/svg+xml MIME types.

Setting	Type	Description

backticks	Int	The number of backticks to enclose raw text markup, which may be increased to disambiguiate nested raw text.

block	Bool	When mode = math, specifies whether the enclosing dollar signs are padded with a space to render the element inline or its own block.

depth	Int	The current level of nesting within container types to specify the degree of indentation by repeating the tab_size.

mode	Mode	The current Typst syntactical context where code follows the number sign, markup is at the top-level and enclosed in square brackets, and math is enclosed in dollar signs.

parenthesize	Bool	Whether to enclose some mathematical elements in parentheses to specify their operator precedence and avoid ambiguity.

preamble	TypstString	Used at the beginning of Typst source files generated by render and some show methods.

tab_size	Int	The number of spaces used by some elements with multi-line Typst formatting, which is repeated for each level of depth

Examples

julia> context
TypstContext with 7 entries:
  :mode         => markup
  :parenthesize => true
  :block        => false
  :preamble     => TypstString(TypstText("#set page(margin: 1em, height: auto, …
  :tab_size     => 2
  :backticks    => 3
  :depth        => 0

Typstry.reset_context — Function

reset_context()

Remove any custom mappings from the context such that it is returned to its default state.

Examples

julia> reset_context()
TypstContext with 7 entries:
  :mode         => markup
  :parenthesize => true
  :block        => false
  :preamble     => TypstString(TypstText("#set page(margin: 1em, height: auto, …
  :tab_size     => 2
  :backticks    => 3
  :depth        => 0