Command

Typstry.TypstCommand — Type

TypstCommand
TypstCommand(::Vector{String})
TypstCommand(::TypstCommand; kwargs...)

The Typst command-line interface.

This command attempts to support the same interface as Cmd. However, this interface is unspecified which may result in missing functionality.

Examples

julia> help = TypstCommand(["help"])
typst`help`

julia> TypstCommand(help; ignorestatus = true)
typst`help`

Typstry.@typst_cmd — Macro

@typst_cmd(parameters)
typst`parameters...`

Construct a TypstCommand without interpolation.

Each parameter must be separated by a space " ".

Examples

julia> typst`help`
typst`help`

julia> typst`compile input.typ output.typ`
typst`compile input.typ output.typ`