Internals

Strings

Typstry.settings — Constant

settings

A constant NamedTuple containing the default IOContext settings for show with the "text/typst" MIME type.

Typstry.code_mode — Function

code_mode(io)

Print the number sign "#" unless mode(io) == code.

See also mode.

Examples

julia> Typstry.code_mode(IOContext(stdout, :mode => code))

julia> Typstry.code_mode(IOContext(stdout, :mode => markup))
#

julia> Typstry.code_mode(IOContext(stdout, :mode => math))
#

Typstry.depth — Function

depth(io)

Return io[depth]::Int64.

Typstry.enclose — Function

enclose(f, io, x, left, right = reverse(left); settings...)

Call f(io, x; settings...) between printing left and right, respectfully.

Examples

julia> Typstry.enclose((io, i; x) -> print(io, i, x), stdout, 1, "\$ "; x = "x")
$ 1x $

Typstry.escape_quote — Function

escape_quote(io, s)

Print the string, with quotes "\"" escaped.

Examples

julia> Typstry.escape_quote(stdout, TypstString("a"))
"\"a\""

Typstry.indent — Function

indent(io)

Return io[indent]::String.

Typstry.inline — Function

inline(io)

Return io[inline]::Bool.

Typstry.join_with — Function

join_with(f, io, xs, delimeter; settings...)

Similar to join, except printing with f(io, x; settings...).

Examples

julia> Typstry.join_with((io, i; x) -> print(io, -i, x), stdout, 1:4, ", "; x = "x")
-1x, -2x, -3x, -4x

Typstry.math_pad — Function

math_pad(io, x)

Return "", "\$", or "\$ " depending on the mode and inline settings.

Examples

julia> Typstry.math_pad(IOContext(stdout, :mode => math))
""

julia> Typstry.math_pad(IOContext(stdout, :mode => markup, :inline => true))
"\$"

julia> Typstry.math_pad(IOContext(stdout, :mode => markup, :inline => false))
"\$ "

Typstry.mode — Function

mode(io)

Return io[mode]::Mode.

Typstry.print_parameters — Function

print_parameters(io, f, keys)

Print the name of a Typst function, an opening parenthesis, the parameters to a Typst function, and a newline.

Skip keys that are not in the IOContext.

Examples

julia> Typstry.print_parameters(
           IOContext(stdout, :delim => "\"(\""),
       "vec", [:delim, :gap])
vec(delim: "(",

Commands

Typstry.typst_program — Constant

typst_program

A constant Cmd that is the Typst command-line interface given by Typst_jll.jl with no additional parameters.