Strings

Typstry.TypstString — Type

TypstString <: AbstractString
TypstString(x; settings...)

Construct a string with print_typst.

Typstry.@typst_str — Macro

@typst_str(s)
typst"s"

Construct a TypstString with custom interpolation and escaping. Backslashes \ and quotation marks " must still be escaped.

The syntax for interpolation is a call to the TypstString constructor, except with a backslash \ instead of the type name.

Tip

Use print_typst to print directly to an IO.

See also the performance tip to avoid string interpolation for I/O.

Examples

julia> x = 1;

julia> typst"$ \(x) / \(x + 1) $"
typst"\$ 1 / 2 \$"

julia> typst"\(x // 2)"
typst"\$ 1 / 2 \$"

julia> typst"\(x // 2; mode = math)"
typst"1 / 2"

julia> typst"\\(x)"
typst"\\\\(x)"

Typstry.Mode — Type

Mode

An Enum to indicate whether the current mode is code, markup, or math.

julia> Mode
Enum Mode:
code = 0
markup = 1
math = 2

Typstry.format — Function

format(io, x; settings...)

Typstry.print_typst — Function

print_typst(io = stdout, x;
    mode = markup,
    inline = false,
    tab = ' ' ^ 4,
    depth = 0,
settings...)

See also Mode.

Interface

Base.:* — Function

*(::TypstString, ::TypstString)

Base.IOBuffer — Type

IOBuffer(::TypstString)

Base.codeunit — Function

codeunit(::TypstString)

codeunit(::TypstString, ::Integer)

Base.isvalid — Function

isvalid(::TypstString, ::Integer)

Base.iterate — Function

iterate(::TypstString)

iterate(::TypstString, ::Integer)

Base.ncodeunits — Function

ncodeunits(::TypstString)

Base.pointer — Function

pointer(::TypstString)

Base.show — Method

show(::IO, ::TypstString)