Strings

Typstry

Typstry.Mode — Type

Mode

An Enumerated type used to specify that the current Typst syntactical context is code, markup, or math.

julia> Mode
Enum Mode:
code = 0
markup = 1
math = 2

Typstry.Typst — Type

Typst{T}
Typst(::T)

A wrapper used to pass values to show(::IO, ::MIME"text/typst", ::Typst).

julia> Typst(1)
Typst{Int64}(1)

julia> Typst("a")
Typst{String}("a")

Typstry.TypstString — Type

TypstString <: AbstractString
TypstString(::Any; context...)

Convert the value to a Typst formatted string.

Optional Julia settings and Typst parameters are passed to show(::IO, ::MIME"text/typst", ::Typst) in an IOContext. See also show_typst for a list of supported types.

Info

This type implements the String interface. However, the interface is unspecified and may result in unexpected behavior.

Examples

julia> TypstString("a")
typst"\"a\""

julia> TypstString("a"; mode = code)
typst"\"\\\"a\\\"\""

Typstry.TypstText — Type

TypstText{T}
TypstText(::Any)

A wrapper whose show_typst method uses print.

Info

This may be used to insert control characters into a TypstString. Unescaped control characters in TypstStrings may break formatting in some environments, such as the REPL.

Examples

julia> TypstText(1)
TypstText{Int64}(1)

julia> TypstText("a")
TypstText{String}("a")

Typstry.@typst_str — Macro

@typst_str(s)
typst"s"

Construct a TypstString.

Control characters are escaped, except double quotation marks and backslashes in the same manner as @raw_str. Values may be interpolated by calling the TypstString constructor, except using a backslash instead of the type name.

Tip

Print directly to an IO using show(::IO, ::MIME"text/typst", ::Typst).

See also the performance tip to Avoid string interpolation for I/O.

Examples

julia> x = 1;

julia> typst"$ \(x) / \(x + 1) $"
typst"$ 1 / 2 $"

julia> typst"\(x // 2)"
typst"$1 / 2$"

julia> typst"\(x // 2; mode = math)"
typst"(1 / 2)"

julia> typst"\\(x)"
typst"\\(x)"

Typstry.code — Constant

code

A Typst syntactical Mode prefixed by the number sign.

Examples

julia> code
code::Mode = 0

Typstry.markup — Constant

markup

A Typst syntactical Mode at the top-level of source text and enclosed within square brackets.

julia> markup
markup::Mode = 1

Typstry.math — Constant

math

A Typst syntactical Mode enclosed within dollar signs.

julia> math
math::Mode = 2

Typstry.context — Function

context(x)

Provide formatting data for show(::IO, ::MIME"text/typst", ::Typst).

Implement this function for a custom type to specify its custom settings and parameters. Passing a value wrapped in Typst will merge! its custom context with defaults, such that the defaults may be overwritten. To be compatible with merging contexts and constructing an IOContext, methods must return an AbstractDict{Symbol}.

See also TypstString.

Setting	Default	Type	Description

:block	false	Bool	When :mode => math, specifies whether the enclosing dollar signs are padded with a space to render the element inline or its own block.

:depth	0	Int	The current level of nesting within container types to specify the degree of indentation.

:indent	' ' ^ 4	String	The string used for horizontal spacing by some elements with multi-line Typst formatting.

:mode	markup	Mode	The current Typst syntactical context where code follows the number sign, markup is at the top-level and enclosed in square brackets, and math is enclosed in dollar signs.

:parenthesize	true	Bool	Whether to enclose some mathematical elements in parentheses to specify their operator precedence and avoid ambiguity.

Typstry.show_typst — Function

show_typst(x)

Print to stdout in Typst format with the default and custom contexts.

Examples

julia> show_typst(1 // 2)
$1 / 2$

julia> show_typst(1:4)
$vec(
    1, 2, 3, 4
)$

show_typst(io, x)

Print in Typst format with Julia settings and Typst parameters provided provided by an IOContext.

Implement this function for a custom type to specify its Typst formatting. Settings are used in Julia to format the TypstString and have varying types. Parameters are passed to a function in the Typst source file and must be a String with the same name as in Typst, except that dashes are replaced with underscores.

For additional information on settings and parameters, see also context and the Typst Documentation, respectively.

Info

... indicates that the method may call show(::IO, ::MIME"text/typst", ::Typst) to format a value which may require additional settings.

Warning

This function's methods are incomplete. Please file an issue or create a pull-request for missing methods.

Type	Settings	Parameters

AbstractArray	:block, :depth, :indent, :mode, ...	:delim, :gap

AbstractChar	:mode

AbstractFloat

AbstractMatrix	:block, :depth, :indent, :mode, ...	:augment, :column_gap, :delim, :gap, :row_gap

AbstractString	:mode

Bool	:mode

Complex{Bool}	:block, :mode, :parenthesize

Complex	:block, :mode, :parenthesize, ...

Irrational	:mode, ...

Nothing	:mode

OrdinalRange{<:Integer, <:Integer}	:mode, ...

Rational	:block, :mode, :parenthesize, ...

Regex	:mode

Signed

StepRangeLen{<:Integer, <:Integer, <:Integer}	:mode, ...

String	:mode

Text	:mode, ...

Tuple	:block, :depth, :indent, :mode, ...	:delim, :gap

Typst	...

TypstString

TypstText

Unsigned	:mode

Examples

julia> show_typst(stdout, 1)
1

julia> show_typst(IOContext(stdout, :mode => code), "a")
"\"a\""

Base

Base.IOBuffer — Type

IOBuffer(::TypstString)

See also TypstString.

Examples

julia> IOBuffer(typst"a")
IOBuffer(data=UInt8[...], readable=true, writable=false, seekable=true, append=false, size=1, maxsize=Inf, ptr=1, mark=-1)

Base.codeunit — Function

codeunit(::TypstString)
codeunit(::TypstString, ::Integer)

See also TypstString.

Examples

julia> codeunit(typst"a")
UInt8

julia> codeunit(typst"a", 1)
0x61

Base.isvalid — Function

isvalid(::TypstString, ::Integer)

See also TypstString.

Examples

julia> isvalid(typst"a", 1)
true

Base.iterate — Method

iterate(::TypstString, ::Integer)
iterate(::TypstString)

See also TypstString.

Examples

julia> iterate(typst"a")
('a', 2)

julia> iterate(typst"a", 1)
('a', 2)

Base.ncodeunits — Function

ncodeunits(::TypstString)

See also TypstString.

Examples

julia> ncodeunits(typst"a")
1

Base.pointer — Function

pointer(::TypstString)

See also TypstString.

Base.repr — Function

repr(::MIME, ::TypstString; kwargs...)

See also TypstString.

Info

This method patches incorrect output from the assumption in repr that the parameter is already in the requested MIME type when the MIME type istextmime and the parameter is an AbstractString.

Examples

julia> repr(MIME"text/plain"(), typst"a")
"typst\"a\""

julia> repr(MIME"text/typst"(), typst"a")
typst"a"

Base.show — Method

show(::IO, ::TypstString)

See also TypstString.

Examples

julia> show(stdout, typst"a")
typst"a"

Base.show — Method

show(::IO, ::MIME"text/typst", ::Union{Typst, TypstString, TypstText})

Print the Typst format.

This method provides formatting data to show_typst specified by a default and custom context.

See also TypstString and TypstText.

Examples

julia> show(stdout, "text/typst", typst"a")
a

julia> show(stdout, "text/typst", Typst("a"))
"a"

julia> show(IOContext(stdout, :mode => code), "text/typst", Typst("a"))
"\"a\""

Base.show — Method

show(::IO, ::Union{
    MIME"application/pdf", MIME"image/png", MIME"image/svg+xml"
}, ::TypstString)

Print the Portable Document Format (PDF), Portable Network Graphics (PNG), or Scalable Vector Graphics (SVG) format using render.

Note

Environments such as Pluto.jl notebooks use this function to display TypstStrings.