Internals

This reference documents non-public utilities.

Typstry.compile_workload — Function

compile_workload(examples)

Given an iterable of value-type pairs, interpolate each value into a @typst_str within a PrecompileTools.@compile_workload block.

Typstry.unwrap — Function

unwrap(io, type, key)
unwrap(io, type, key, default)

Strings

Typstry.Strings — Module

Strings

The Julia to Typst interface, a custom string corresponding to Typst source text, and its implementation of the String interface.

Examples

julia> Typstry.Strings
Typstry.Strings

Typstry.Strings.examples — Constant

examples

A constant Vector of Julia values and their corresponding Types implemented for show_typst.

Typstry.Strings.typst_mime — Constant

typst_mime

Equivalent to MIME"text/typst"().

Examples

julia> Typstry.Strings.typst_mime
MIME type text/typst

Typstry.Strings.backticks — Function

backticks(io)

Examples

julia> Typstry.Strings.backticks(IOContext(stdout, :backticks => 3))
3

Typstry.Strings.block — Function

block(io)

Examples

julia> Typstry.Strings.block(IOContext(stdout, :block => true))
true

Typstry.Strings.code_mode — Function

code_mode(io)

Print the number sign, unless mode(io) == code.

See also Mode and mode.

Examples

julia> Typstry.Strings.code_mode(IOContext(stdout, :mode => code))

julia> Typstry.Strings.code_mode(IOContext(stdout, :mode => markup))
#

julia> Typstry.Strings.code_mode(IOContext(stdout, :mode => math))
#

Typstry.Strings.depth — Function

depth(io)

Examples

julia> Typstry.Strings.depth(IOContext(stdout, :depth => 0))
0

Typstry.Strings.enclose — Function

enclose(f, io, x, left, right = reverse(left); kwargs...)

Call f(io, x; kwargs...) between printing left and right, respectfully.

Examples

julia> Typstry.Strings.enclose((io, i; x) -> print(io, i, x), stdout, 1, "\$ "; x = "x")
$ 1x $

Typstry.Strings.escape — Function

escape(io, n)

Print \ to io n times.

Examples

julia> Typstry.Strings.escape(stdout, 2)
\\

Typstry.Strings.indent — Function

indent(io)

Examples

julia> Typstry.Strings.indent(IOContext(stdout, :tab_size => 2))
"  "

Typstry.Strings.join_with — Function

join_with(f, io, xs, delimeter; kwargs...)

Similar to join, except printing with f(io, x; kwargs...).

Examples

julia> Typstry.Strings.join_with((io, i; x) -> print(io, -i, x), stdout, 1:4, ", "; x = "x")
-1x, -2x, -3x, -4x

Typstry.Strings.math_mode — Function

math_mode(f, io, x; kwargs...)

Typstry.Strings.math_pad — Function

math_pad(io)

Return "", "\$", or "\$ " depending on the block and mode settings.

Examples

julia> Typstry.Strings.math_pad(IOContext(stdout, :mode => math))
""

julia> Typstry.Strings.math_pad(IOContext(stdout, :block => true, :mode => markup))
"\$ "

julia> Typstry.Strings.math_pad(IOContext(stdout, :block => false, :mode => markup))
"\$"

Typstry.Strings.mode — Function

mode(io)

Examples

julia> Typstry.Strings.mode(IOContext(stdout, :mode => code))
code::Mode = 0

Typstry.Strings.parenthesize — Function

parenthesize(io)

Examples

julia> Typstry.Strings.parenthesize(IOContext(stdout, :parenthesize => true))
true

Typstry.Strings.show_array — Function

show_array(io, x)

Typstry.Strings.show_parameters — Function

show_parameters(io, f, keys, final)

Examples

julia> Typstry.Strings.show_parameters(
           IOContext(stdout, :depth => 0, :tab_size => 2, :delim => typst"\"(\""),
       "vec", [:delim, :gap], true)
vec(
  delim: "(",

Typstry.Strings.show_raw — Function

show_raw(f, io, x, language)

Typstry.Strings.show_vector — Function

show_vector(io, x)

Dates.jl

Info

A Dates.jl package extension would currently print warnings during precompilation. See also the Julia issue #52511

Typstry.Strings.date_time — Function

date_time(::Union{Dates.Date, Dates.Time, Dates.DateTime})

Typstry.Strings.duration — Function

duration(::Dates.Period)

Examples

julia> Typstry.Strings.duration(Dates.Day(1))
:days

julia> Typstry.Strings.duration(Dates.Hour(1))
:hours

Typstry.Strings.dates — Function

dates(::Union{Dates.Date, Dates.DateTime, Dates.Period, Dates.Time})

Examples

julia> Typstry.Strings.dates(Dates.Date(1))
("datetime", (:year, :month, :day), (1, 1, 1))

julia> Typstry.Strings.dates(Dates.Day(1))
("duration", (:days,), (TypstText{String}("1"),))

Commands

Typstry.Commands — Module

Commands

A custom command corresponding to the Typst compiler and its implementation of the Cmd interface.

Examples

julia> Typstry.Commands
Typstry.Commands

Typstry.Commands.default_preamble — Constant

default_preamble

Typstry.Commands.apply — Function

apply(f, tc, args...; kwargs...)

Typstry.Commands.format — Function

format(::Union{MIME"application/pdf", MIME"image/png", MIME"image/svg+xml"})

Return the image format acronym corresponding to the given MIME.

Examples

julia> Typstry.Commands.format(MIME"application/pdf"())
"pdf"

julia> Typstry.Commands.format(MIME"image/png"())
"png"

julia> Typstry.Commands.format(MIME"image/svg+xml"())
"svg"