Utilities

Typstry.Mode — Type

Mode

An Enumerated type used to specify that the current Typst syntactical context is code, markup, or math.

Examples

julia> Mode
Enum Mode:
code = 0
markup = 1
math = 2

Typstry.code — Constant

code

A Typst syntactical Mode prefixed by the number sign.

Examples

julia> code
code::Mode = 0

Typstry.markup — Constant

markup

A Typst syntactical Mode at the top-level of source text and enclosed within square brackets.

julia> markup
markup::Mode = 1

Typstry.math — Constant

math

A Typst syntactical Mode enclosed within dollar signs.

julia> math
math::Mode = 2

Typstry.ContextError — Type

ContextError <: Exception
ContextError(::Type, ::Type, ::Symbol)

An Exception indicating that a context key returned a value of an incorrect type.

Interface

Implements the Exception interface.

showerror(::IO, ::ContextError)

show(::IO, ::MIME"text/plain", ::ContextError)

Examples

julia> ContextError(Mode, String, :mode)
ContextError(Mode, String, :mode)

Typstry.TypstContext — Type

TypstContext <: AbstractDict{Symbol, Any}
TypstContext(::Any)
TypstContext(; kwargs...)

Provide formatting data for show_typst.

Implement a method of this constructor for a custom type to specify its custom settings and parameters.

Calls to show_typst from the following methods:

TypstString

render

show_typst(::IO, ::TypstContext, x)

show with the application/pdf, image/png, image/svg+xml, and text/typst MIME types and a TypstString, TypstText, and Typst value

specify the TypstContext by combining the following contexts:

A default context

The context specified by set_context, which is combined with the context upon initialization

The TypstContext constructor implemented for the given type, the context specified in the call by keyword parameters, a given TypstContext, or the IOContext key :typst_context, depending on the calling method

Any context specified by a recursive call in show_typst to format values, such as elements from a container

Duplicate keys are handled such that each successive context overwrites those of previous contexts, prioritized in order as listed. In other words, the default context has the lowest priority while recursive calls to show_typst have the highest priority.

Interfaces

This type implements the dictionary and iteration interfaces. However, it is immutable such that it does not support inserting, deleting, or setting a key-value pair.

eltype(::TypstContext)

get(::TypstContext, ::Symbol, default)

get(::Union{Function, Type}, ::TypstContext, ::Symbol)

iterate(::TypstContext, state)

iterate(::TypstContext)

length(::TypstContext)

show(::IO, ::TypstContext)

Typstry.context — Constant

context

A constant TypstContext used default formatting data when calling show_typst.

May be configured using set_context.

Setting	Type	Description

backticks	Int	The number of backticks to enclose raw text markup, which may be increased to disambiguiate nested raw text.

block	Bool	When mode = math, specifies whether the enclosing dollar signs are padded with a space to render the element inline or its own block.

depth	Int	The current level of nesting within container types to specify the degree of indentation by repeating the tab_size.

mode	Mode	The current Typst syntactical context where code follows the number sign, markup is at the top-level and enclosed in square brackets, and math is enclosed in dollar signs.

parenthesize	Bool	Whether to enclose some mathematical elements in parentheses to specify their operator precedence and avoid ambiguity.

preamble	TypstString	Used at the beginning of Typst source files generated by render and some show methods.

tab_size	Int	The number of spaces used by some elements with multi-line Typst formatting, which is repeated for each level of depth

Examples

julia> context
TypstContext with 7 entries:
  :mode         => markup
  :parenthesize => true
  :block        => false
  :preamble     => TypstString(TypstText("#set page(margin: 1em, height: auto, …
  :tab_size     => 2
  :backticks    => 3
  :depth        => 0

Typstry.set_context — Function

set_context(::TypstContext)
set_context()

Use Preferences.jl such that after restarting Julia, the context is initialized to the given TypstContext merged with default settings.

Specifying a key contained in the default settings will override it. If a TypstContext is not provided, the context is reset to the default settings.

Tip

Use this function to customize the default formatting in environments that display values using show with the application/pdf, image/png, and image/svg+xml MIME types.