Manual

Interface

Typstry.@typst_str — Macro

@typst_str(s)

Construct a string with custom interpolation and without unescaping. Backslashes (\) and quotation marks (") must still be escaped.

Use $$ to interpolate values. The $ symbol is not used because Typst uses that symbol to start and end a math mode block.

Examples

julia> x = 1;

julia> typst"$1 / x$"
"\$1 / x\$"

julia> typst"$$x $$(x + 1)"
"1 2"

julia> typst"\$$x"
"\$\$x"

julia> typst"\\"
"\\"

Typstry.@typst_cmd — Macro

@typst_cmd(s)

Return a Cmd such that run(typst`$s`) is equivalent to shell> typst $s.

Note

This macro does not yet support interpolation.

Examples

julia> run(typst`help`);
The Typst compiler

Usage: typst [OPTIONS] <COMMAND>

Commands:
  compile  Compiles an input file into a supported output format [aliases: c]
  watch    Watches an input file and recompiles on changes [aliases: w]
  query    Processes an input file to extract provided metadata
  fonts    Lists all discovered fonts in system and custom font paths
  update   Self update the Typst CLI (disabled)
  help     Print this message or the help of the given subcommand(s)

Options:
  -v, --verbosity...  Sets the level of logging verbosity: -v = warning & error, -vv = info, -vvv = debug, -vvvv = trace
      --cert <CERT>   Path to a custom CA certificate to use when making network requests [env: TYPST_CERT=]
  -h, --help          Print help
  -V, --version       Print version

julia> write("input.typ", typst"$x ^ 2$");

julia> run(typst`compile input.typ output.pdf`);

Typstry.typst — Function

typst(s)

Equivalent to shell> typst $s.

Examples

julia> typst("help");
The Typst compiler

Usage: typst [OPTIONS] <COMMAND>

Commands:
  compile  Compiles an input file into a supported output format [aliases: c]
  watch    Watches an input file and recompiles on changes [aliases: w]
  query    Processes an input file to extract provided metadata
  fonts    Lists all discovered fonts in system and custom font paths
  update   Self update the Typst CLI (disabled)
  help     Print this message or the help of the given subcommand(s)

Options:
  -v, --verbosity...  Sets the level of logging verbosity: -v = warning & error, -vv = info, -vvv = debug, -vvvv = trace
      --cert <CERT>   Path to a custom CA certificate to use when making network requests [env: TYPST_CERT=]
  -h, --help          Print help
  -V, --version       Print version

julia> write("input.typ", typst"$x ^ 2$");

julia> typst("compile input.typ output.pdf");

typst(xs...)

Equivalent to typst(join(xs, " ")).

Examples

julia> write("input.typ", typst"$x ^ 2$");

julia> typst("compile", "input.typ", "output.pdf");

Typstry.render — Function

render(::String; input = "input.typ", output = "output.pdf", open = true)

Render the given string to a document.

This function generates two files. The first is the input, which contains the Typst document. The second is the output, which is rendered from the input using Typst's compile function.

The document format is inferred by the file extension of output. The available formats are pdf, png, and svg.

If open, output will be opened using the default viewer.

Note

This is designed to generate a document with little effort. For more advanced useage, see typst.

Examples

julia> render(typst"$x ^ 2$");

julia> render([1, 2, 3, 4]);

render(elements...; delimeter = " \\n", kwargs...)

Equivalent to render(join(xs, delimeter; kwargs...)).

Examples

julia> render("The area of a circle is", typst"$A = pi dot r ^ 2$"; delimeter = " ");

Internals

Typstry.interpolate — Function

interpolate(xs, s, start, stop)

Typstry.typst_cmd — Function

typst_cmd(s)