Typstry.compile — Function

compile(input, [output]; pre_options = (), post_options = ())

Equivalent to typst(compile, input, output; pre_options, post_options). ```

Typstry.fonts — Method

fonts(; pre_options = (), post_options = ())

Equivalent to typst(fonts; pre_options, post_options).

Typstry.help — Method

help([command::Function])

Prints the help message for typst or the given command.

Equivalent to typst(help, command).

Examples

julia> help();
typst creates PDF files from .typ files

Usage: typst [OPTIONS] <COMMAND>

Commands:
  compile  Compiles the input file into a PDF file [aliases: c]
  watch    Watches the input file and recompiles on changes [aliases: w]
  fonts    List all discovered fonts in system and custom font paths
  help     Print this message or the help of the given subcommand(s)

Options:
      --font-path <DIR>  Add additional directories to search for fonts [env: TYPST_FONT_PATHS=]
      --root <DIR>       Configure the root for absolute paths [env: TYPST_ROOT=]
  -v, --verbosity...     Sets the level of verbosity: 0 = none, 1 = warning & error, 2 = info, 3 = debug, 4 = trace
  -h, --help             Print help
  -V, --version          Print version

julia> help(fonts);
List all discovered fonts in system and custom font paths

Usage: typst fonts [OPTIONS]

Options:
      --variants  Also list style variants of each font family
  -h, --help      Print help

Typstry.typst — Method

typst(args...)

Call the typst command-line interface with the given arguments via Typst_jll.jl.

Equivalent to $ typst args....

Typstry.typst — Method

typst(command::Function, args...; pre_options = (), post_options = ())

Equivalent to typst(pre_options..., string(command), post_options..., args...).

Typstry.watch — Function

watch(input, [output]; pre_options = (), post_options = ())

Equivalent to typst(watch, input, output; pre_options, post_options).

Typstry.@t_str — Macro

t_str(s)

Construct a string without interpolation and unescaping (except for quotation marks, ", which still have to be escaped).

Note

This macro is currently equivalent to @raw_str. Future work will implement custom interpolation.

Examples

julia> t"$1 / x$"
"\$1 / x\$"