Command

Typstry.TypstCommand — Type

TypstCommand
TypstCommand(::Vector{String})
TypstCommand(::TypstCommand; kwargs...)

Typstry.@typst_cmd — Macro

@typst_cmd(parameters)
typst`parameters...`

Return a TypstCommand.

Examples

julia> typst`help`
typst`help`

julia> typst`compile input.typ output.typ`
typst`compile input.typ output.typ`

Typstry.render — Function

render(elements...;
    delimeter = '\n',
    input = "input.typ",
    output = "output.pdf",
    open = true,
settings...)

Render the elements to a document.

Each element is written to the input file with print_typst and the given settings, seperated by the delimeter.

Then, the input file is compiled to the output file with a TypstCommand.

The document format is inferred by the file extension of output. The available formats are pdf, png, and svg.

If open = true, theoutput` file will be opened using the default viewer.

Examples

julia> render(typst"$x ^ 2$");

julia> render([1 2; 3 4]);

Interface

Base.show — Method

show(::IO, ::TypstCommand)

Base.run — Function

run(::TypstCommand, args...; kwargs...)

Base.addenv — Function

addenv(::TypstCommand, args...; kwargs...)

Base.setenv — Function

setenv(::TypstCommand, env; kwargs...)

Base.ignorestatus — Function

ignorestatus(::TypstCommand)

Base.detach — Function

detach(::TypstCommand)

Base.setcpuaffinity — Function

setcpuaffinity(::TypstCommand, cpus)