Commands

Typstry

Typstry.TypstCommand — Type

TypstCommand(::Vector{String})
TypstCommand(::TypstCommand; kwargs...)

The Typst command-line interface.

Info

This type implements the Cmd interface. However, the interface is unspecified which may result unexpected behavior.

Examples

julia> help = TypstCommand(["help"])
typst`help`

julia> TypstCommand(help; ignorestatus = true)
typst`help`

Typstry.@typst_cmd — Macro

@typst_cmd(parameters)
typst`parameters...`

Construct a TypstCommand without interpolation.

Each parameter must be separated by a space " ".

Examples

julia> typst`help`
typst`help`

julia> typst`compile input.typ output.typ`
typst`compile input.typ output.typ`

Typstry.julia_mono — Constant

julia_mono

An Artifact containing the JuliaMono typeface.

Base

Base.addenv — Function

addenv(::TypstCommand, args...; kwargs...)

See also TypstCommand.

Examples

julia> addenv(typst`compile input.typ output.pdf`, "TYPST_FONT_PATHS" => julia_mono)
typst`compile input.typ output.pdf`

Base.detach — Function

detach(::TypstCommand)

See also TypstCommand.

Examples

julia> detach(typst`help`)
typst`help`

Base.ignorestatus — Function

ignorestatus(::TypstCommand)

See also TypstCommand.

Examples

julia> ignorestatus(typst`help`)
typst`help`

Base.run — Function

run(::TypstCommand, args...; kwargs...)

See also TypstCommand.

Base.setcpuaffinity — Function

setcpuaffinity(::TypstCommand, cpus)

See also TypstCommand.

Compat

Requires at least Julia v0.8.

Examples

julia> setcpuaffinity(typst`help`, [1])
typst`help`

Base.setenv — Function

setenv(::TypstCommand, env; kwargs...)

See also TypstCommand.

Base.show — Method

show(::IO, ::TypstCommand)

See also TypstCommand.

Examples

julia> show(stdout, typst`help`)
typst`help`