Internals

Strings

Typstry.examples — Constant

examples

A constant Vector{Pair{Any, Type}} where the first element is a value and the second element is the (potentially abstract) type corresponding to its show(::IO, ::MIME"text/plain", ::Any) method.

Typstry.preamble — Constant

preamble

Typstry.settings — Constant

settings

A constant NamedTuple containing the default IOContext settings for show(::IO, ::MIME"text/typst", ::Any).

Typstry.typst_mime — Constant

typst_mime

A constant equal to MIME"text/typst"().

Examples

julia> Typstry.typst_mime
MIME type text/typst

Typstry.code_mode — Function

code_mode(io)

Print the number sign # unless mode(io) == code.

See also Mode and mode.

Examples

julia> Typstry.code_mode(IOContext(stdout, :mode => code))

julia> Typstry.code_mode(IOContext(stdout, :mode => markup))
#

julia> Typstry.code_mode(IOContext(stdout, :mode => math))
#

Typstry.depth — Function

depth(io)

Return io[:depth]::Int

Examples

julia> Typstry.depth(IOContext(stdout, :depth => 0))
0

Typstry.enclose — Function

enclose(f, io, x, left, right = reverse(left); settings...)

Call f(io, x; settings...) between printing left and right, respectfully.

Examples

julia> Typstry.enclose((io, i; x) -> print(io, i, x), stdout, 1, "\$ "; x = "x")
$ 1x $

Typstry.escape_quote — Function

escape_quote(io, s)

Print the string, with quotes \ escaped.

Examples

julia> Typstry.escape_quote(stdout, TypstString("a"))
"\"a\""

Typstry.indent — Function

indent(io)

Return io[:indent]::String.

Examples

julia> Typstry.indent(IOContext(stdout, :indent => ' ' ^ 4))
"    "

Typstry.inline — Function

inline(io)

Return io[:inline]::Bool.

Examples

julia> Typstry.inline(IOContext(stdout, :inline => true))
true

Typstry.join_with — Function

join_with(f, io, xs, delimeter; settings...)

Similar to join, except printing with f(io, x; settings...).

Examples

julia> Typstry.join_with((io, i; x) -> print(io, -i, x), stdout, 1:4, ", "; x = "x")
-1x, -2x, -3x, -4x

Typstry.math_pad — Function

math_pad(io, x)

Return "", "\$", or "\$ " depending on the mode and inline settings.

Examples

julia> Typstry.math_pad(IOContext(stdout, :mode => math))
""

julia> Typstry.math_pad(IOContext(stdout, :mode => markup, :inline => true))
"\$"

julia> Typstry.math_pad(IOContext(stdout, :mode => markup, :inline => false))
"\$ "

Typstry.mode — Function

mode(io)

Return io[:mode]::Mode.

See also [Mode].

Examples

julia> Typstry.mode(IOContext(stdout, :mode => code))
code::Mode = 0

Typstry.print_parameters — Function

print_parameters(io, f, keys)

Print the name of a Typst function, an opening parenthesis, the parameters to a Typst function, and a newline.

Skip keys that are not in the IOContext.

Examples

julia> Typstry.print_parameters(
           IOContext(stdout, :delim => "\"(\""),
       "vec", [:delim, :gap])
vec(delim: "(",

Typstry.show_image — Function

show_image(io, format, t)

Typstry.static_parse — Function

static_parse(args...; filename, kwargs...)

Call Meta.parse with the filename if it is supported in the current Julia version (at least v1.10).

Commands

Typstry.typst_program — Constant

typst_program

A constant Cmd that is the Typst command-line interface given by Typst_jll.jl with no additional parameters.

Typstry.apply — Function

apply(f, tc, args...; kwargs...)